Move FileSelectMenu and FileSelectMenuItem to src/components/files
