+++ src/components/instruments/instrument-settings.tsx
@@ -1,6 +1,6 @@
 import { ConfirmButton } from "components/confirm-button";
 import { ErrorAlert } from "components/error-alert";
-import { FileSelectMenu } from "components/file-select-menu";
+import { FileSelectMenu } from "components/files/file-select-menu";
 import { SelectMenu, SelectMenuItem } from "components/select-menu";
 import { FormField } from "components/forms/form-field";
 import { NoteSelectMenu } from "components/note-select-menu";
